pyproject.toml
[project]
name = "townlet"
version = "0.1.0"
description = "A pedagogical Deep Reinforcement Learning environment designed to trick students into learning graduate-level RL"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "John", email = "john@foundryside.dev" }
]
keywords = [
    "reinforcement-learning",
    "deep-learning",
    "machine-learning",
    "ai",
    "education",
    "drl",
    "q-learning",
    "lstm",
    "pomdp",
    "curriculum-learning",
    "intrinsic-motivation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Framework :: Pytest",
    "Typing :: Typed",
]
dependencies = [
    "pettingzoo>=1.24.0",
    "gymnasium>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.9.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "cloudpickle>=3.0.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "mlflow>=2.9.0",
    "tensorflow[and-cuda]>=2.20.0",
    "tensorflow>=2.20.0",
    "msgpack>=1.1.2",
    "lz4>=4.4.5",
]

[project.urls]
Homepage = "https://github.com/tachyon-beep/hamlet"
Repository = "https://github.com/tachyon-beep/hamlet"
Documentation = "https://github.com/tachyon-beep/hamlet/tree/main/docs"
Issues = "https://github.com/tachyon-beep/hamlet/issues"
Changelog = "https://github.com/tachyon-beep/hamlet/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "types-PyYAML>=6.0.12.12",
    "hypothesis>=6.100.0",
    "vulture>=2.0.0",
]
recording = [
    "msgpack>=1.0.0",       # Serialization
    "lz4>=4.0.0",           # Compression
    "ffmpeg-python>=0.2.0", # Video export
    "pillow>=10.0.0",       # Image manipulation
    "matplotlib>=3.7.0",    # Frame rendering
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/townlet"]

[tool.black]
line-length = 140
target-version = ['py313']

[tool.ruff]
line-length = 140
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, can tighten later

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=townlet",
    "--cov-branch",
    "--cov-report=term-missing",
    "-m",
    "not slow",
]
markers = [
    "slow: marks tests as slow (tests that take >5 seconds)",
    "gpu: marks tests requiring CUDA/GPU hardware",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
